--- data/src/Couch/Identified.ts
@@ -1,20 +1,0 @@
-import { JSONSchemaType } from "ajv";
-import { generateSchema } from "../validator";
-
-/**
- * Any object identified by an `id` string.
- */
-export type Identified = {
-  id: string;
-  _rev?: string;
-};
-
-export const { inline, schema, validate } = generateSchema<Identified>({
-  $id: "/couch/identified",
-  type: "object",
-  properties: {
-    id: { type: "string" },
-    _rev: { type: "string", nullable: true },
-  },
-  required: ["id"],
-} as JSONSchemaType<Identified>);
