--- data/src/Util.ts
@@ -1,4 +1,0 @@
-export * as FileRef from "./Util/FileRef";
-export * as ImageRef from "./Util/ImageRef";
-export * as ProcessUpdate from "./Util/ProcessUpdate";
-export * as Text from "./Util/Text";
