+++ data/src/access/PdfManifest.ts
@@ -1,8 +1,8 @@
 import { JSONSchemaType } from "ajv";
-import { Manifest, schema as manifestSchema } from "../Manifest";
-import { Text, inline as textSchema } from "../../Util/Text";
-import { FileRef, inline as fileSchema } from "../../Util/FileRef";
-import { inherit } from "../../validator";
+import { Manifest, schema as manifestSchema } from "./Manifest";
+import { Text, inline as textSchema } from "../util/Text";
+import { FileRef, inline as fileSchema } from "../util/FileRef";
+import { inherit } from "../validator";
 
 type PdfSpec = {
   from: "pdf";
