+++ data/src/access/index.ts
@@ -1,0 +1,5 @@
+export * as Alias from "./Alias";
+export * as Canvas from "./Canvas";
+export * as CanvasManifest from "./CanvasManifest";
+export * as Collection from "./Collection";
+export * as PdfManifest from "./PdfManifest";
