+++ data/src/couch/canvas.spec.ts
@@ -1,9 +1,9 @@
 import test from "ava";
 
-import { toCouch } from "../Couch";
-import { validate } from "./Canvas";
+import { toCouch } from "./util";
+import { validate } from "./canvas";
 
-import { testCanvas } from "../Access/Canvas.spec";
+import { testCanvas } from "../access/Canvas.spec";
 
 test("Canvas couch schema validates a canvas document", (t) => {
   const valid = validate(toCouch(testCanvas));
