+++ data/src/couch/index.ts
@@ -1,0 +1,3 @@
+export * as access from "./access";
+export * as canvas from "./canvas";
+export * as util from "./util";
