--- data/src/mod.ts
@@ -1,3 +1,0 @@
-export * as Access from "./Access";
-
-export { validator } from "./validator";
