+++ data/src/util/ProcessUpdate.ts
@@ -1,6 +1,6 @@
 import { JSONSchemaType } from "ajv";
 import { generateSchema } from "../validator";
-import { Timestamp, inline as timestampSchema } from "../Format/Timestamp";
+import { Timestamp, inline as timestampSchema } from "../format/Timestamp";
 
 /**
  * An object that describes a request for and the output of an automated process that is applied to the parent access object.
