+++ data/src/util/Text.ts
